feat: Use u64 instead of usize for index parameters.
U64 is more appropriate for data that should be the same across
multiple platforms. `usize` is platform specific.

This does not change the use of `NonZeroUsize` for the rolling
window size in virtual storage nodes. Ultimately this values is
used to size an array and that maximum size will be platform
dependent.

Currently there is some casting to `usize`. A separate update
could warn on the `cast_lossless` lint and remove the use of
`as` throughout.

One outstanding query is whether this should actually be i64
to support negative values?

Issue #291.

+++ pywr-core/src/parameters/array.rs
@@ -63,18 +63,18 @@ impl SimpleParameter<f64> for Array1Parameter<f64> {
     }
 }
 
-impl SimpleParameter<usize> for Array1Parameter<u64> {
+impl SimpleParameter<u64> for Array1Parameter<u64> {
     fn compute(
         &self,
         timestep: &Timestep,
         _scenario_index: &ScenarioIndex,
         _values: &SimpleParameterValues,
         _internal_state: &mut Option<Box<dyn ParameterState>>,
-    ) -> Result<usize, PywrError> {
+    ) -> Result<u64, PywrError> {
         let idx = self.timestep_index(timestep);
         // This panics if out-of-bounds
         let value = self.array[[idx]];
-        Ok(value as usize)
+        Ok(value)
     }
 
     fn as_parameter(&self) -> &dyn Parameter
@@ -153,19 +153,19 @@ impl SimpleParameter<f64> for Array2Parameter<f64> {
     }
 }
 
-impl SimpleParameter<usize> for Array2Parameter<u64> {
+impl SimpleParameter<u64> for Array2Parameter<u64> {
     fn compute(
         &self,
         timestep: &Timestep,
         scenario_index: &ScenarioIndex,
         _values: &SimpleParameterValues,
         _internal_state: &mut Option<Box<dyn ParameterState>>,
-    ) -> Result<usize, PywrError> {
+    ) -> Result<u64, PywrError> {
         // This panics if out-of-bounds
         let t_idx = self.timestep_index(timestep);
         let s_idx = scenario_index.indices[self.scenario_group_index];
 
-        Ok(self.array[[t_idx, s_idx]] as usize)
+        Ok(self.array[[t_idx, s_idx]])
     }
 
     fn as_parameter(&self) -> &dyn Parameter
